Improve sending web link with Bluetooth
When sending web link from Google maps app, link cannot be used by
receiver as maps app will send extra text in addition to web link.

Improve the creation of html so that only the web link is
interpreted as a link.

Code is partly copied from MessageView class in the email app.

Change-Id: Idb093cc3d5af5fa6b85649d194fa6dc0c6fdfd1e